add dynamic symbol and decimal for ERC20s
